Define Common SoundManager
